// src/api/stetsom/model/banner.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { BannerStatus } from "./bannerStatus";
import type { I18nString } from "./i18nString";

export interface Banner {
  id: string;
  name: string;
  product_id?: string;
  desktop_image_library_id: string;
  mobile_image_library_id?: string;
  title?: I18nString;
  label?: string;
  href?: string;
  link_url?: string;
  status: BannerStatus;
  available_locales: string[];
  display_from?: string;
  display_until?: string;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order: number;
  created_at: string;
  updated_at: string;
  created_by: string;
}